+++ src/head-sidebar/index.js
@@ -1,0 +1,6 @@
+import { registerPlugin } from '@wordpress/plugins';
+import { HeadPanel } from './head-panel';
+
+registerPlugin( 'talking-head-head-sidebar', {
+	render: HeadPanel,
+} );